requirements.txt
streamlit
openai
requests
pandas
ta
Pillow